Various int -> size_t fixes

+++ src/change.c
@@ -159,7 +159,7 @@ void end_change_chain(void)
 
 static void fix_cursors(size_t offset, size_t del, size_t ins)
 {
-    for (int i = 0; i < buffer->views.count; i++) {
+    for (size_t i = 0; i < buffer->views.count; i++) {
         View *v = buffer->views.ptrs[i];
 
         if (v != view && offset < v->saved_cursor_offset) {
