+++ src/config.c
@@ -24,9 +24,9 @@ static const char *const reset_colors_rc =
 "hi activetab bold\n"
 "hi inactivetab black gray\n";
 
-static bool is_command(const char *str, int len)
+static bool is_command(const char *str, size_t len)
 {
-    for (int i = 0; i < len; i++) {
+    for (size_t i = 0; i < len; i++) {
         if (str[i] == '#') {
             return false;
         }
