+++ src/history.c
@@ -12,7 +12,7 @@ void history_add(PointerArray *history, const char *text, int max_entries)
     }
 
     // Don't add identical entries
-    for (int i = 0; i < history->count; i++) {
+    for (size_t i = 0; i < history->count; i++) {
         if (streq(history->ptrs[i], text)) {
             // Move identical entry to end
             ptr_array_add(history, ptr_array_remove_idx(history, i));
